fix deadlock
The deadlock was arising when readConfig() was failing due to a broken
ssh-agent that was exiting without any errors and non-zero exit codes.

It was only the first part. The second part is that another job was
waiting for condition to be satisfied while there is a case when it
can't be satisfied.