overwrite ENTRYPOINT and CMD
There are images that have a specific entrypoint which makes it
impossible to use this image for CI.

On the other hand, null-ing the entrypoint can cause an error related to
the fact that these images can have empty CMD field.

Therefore, we nullify ENTRYPOINT and set the simplest CMD as possible.